Upstream commits 8071bafb83-67b84dc154
Change CI Docker image

It will allow easily testing supereight in multiple Ubuntu versions.

Fix CMake bugs

Fix the -b option

Fix --image-downsampling-factor

Remove unused command line options

Small code style improvements

- Reorder headers
- Increase spacing
- Fix indentation

Simplify error handling in command line options

Add --help/-h/-? command line option

Make variables static and constexpr

Fix errors in help text

Update code style

Fix returning uninitialized memory

Keep Sophus only where needed

Make se::Image variables non-const

This allows copying one image into another and overwriting the original.
This can be useful for functions with se::Image input parameters that
might need to be initialized or resized. The public interface is still
exactly the same, it is not possible to resize an image after creation.

Add wrapper for std::filesystem

Detects its existence and includes the appropriate header.

Namespace image_utils

Rename se::to_rgba() to se::pack_rgba()

Add RGB/RGBA image conversion functions to image_utils

Fix crash when calling DenseSLAMSystem destructor

The pointer stored in render_T_MC_ is typically not allocated on the
heap. The pipeline should make no assumptions about this and leave
handling the liftime of the passed pointer to the caller.

Reorder command line option declarations

Change fps parameter to float

No need to limit to integer fps values.

Fix mesh saving

Fix by Nils Funk

Fix C++17 filesystem support

GCC will not make the <filesystem> header available unless the C++
standard is explicitly set to C++17.

Change projective functor parameter and member to references

No need to make a copy.

Remove unused variables and functions

Change getMap() signature

The new signature allows getting a const pointer to the map.

Update computeIntegrationScale documentation

Unsure how to document the min_scale parameter.

Remove function parameter errors to suppress compilation warnings

Commented out the variable names in places where they can't be found
elsewhere (e.g. the function declaration).

Fix variable name

Remove unused but set variables

Remove unused function parameters

Fix unsigned/signed int comparison compiler warning

Regarding the variable type changes in kernel_image, I don't think we'll
need image dimensions that need a size_t to fit in the foreseeable
future. Same for the image_levels_.

Comment out unused variable

Not sure if it will be useful in the future

Remove inline specifier from variable

Requires C++17.

Fix uninitialized variable

The name aliasing was causing the warning. The variable outside the if
statement is not truncated so sdf_value is probably a more appropriate
name.

Fix uninitialized variable warning

Fix missing namespace

Fix wrong assert parameters

Add const versions of Node::child()

Make Node member functions const

Fix sensor initialization in se_voxel_impl test

Rename Node::data() to Node::setData() where appropriate

It makes it easier for both the programmer and the compiler to
differentiate between the data() that gets and the data() that sets.
This is because VoxelData data(const Eigen::Vector3i&, const int);
can match void data(const Eigen::Vector3i&, const VoxelData&); if
VoxelData is a POD (int, float, etc.).

Remove const qualifier from Octree::root()

It has no use where it currently is, it just causes a compiler warning.

Add 3DV 2019 reference to README

Add naming conventions

Co-authored with Nils Funk.

Add coordinate frame documentation

Disable raycasting when not needed

Raycasting will be performed only when tracking and/or rendering (GUI)
are enabled.

Initialize se::Octree private variables

Fix memory leak when calling Octree::init() more than once

Set default build type and warn if not Release

Code stolen from OKVIS
https://github.com/ethz-asl/okvis/blob/master/CMakeLists.txt

Put different build types in separate folders

Allows switching between build types without having to recompile
everything.

Fix benchmark script crash

Fix current pose showing in benchmark output

Fix negative timings due to uninitialized memory

Fix rendering issue for MultiresOFusion

Fix by Nils Funk

Merge benchmark into main

This reduces the number of executables from 2*4*3=24 to 2*4*2=16.

The benchmark functionality of showing stats is now available when
passing --no-gui. The previous function of showing just the frame and
camera position has been removed as this information is present in the
new output.

As a temporary measure until we refactor the python benchmark scripts,
the multi-column output of --no-gui is converted to the 4-column output
that the scripts expect.

Cleanup se_apps code

- Rename variables
- Change indentation and spacing
- Make variables const

Add RAM usage to stats shown when --no-gui is used

Add python script to plot supereight stats from --no-gui

Fix README errors and add information

Fix uninitialized variable warning

Fix CMake syntax warning

Fix memory leak in main

When you allocate with new[] deallocate with delete[], not delete.

Fix memory leak in se::Octree

Changed se::Octree::keys_at_depth_ from raw pointer to std::vector.

Set se::Image::data_ size at construction

Previously in the se::Image constructor, memory for data_ was allocated
twice, once in the default std::vector constructor and once more in the
resize. Although most compilers probably optimize this case and only
allocate once, it doesn't hurt to make it more explicit.

Disable TICK/TOCK macros by default

Add missing brackets to improve code style

Initialize camera parameters in config

Restrict pipelines to master and devel branches

Show the current build type instead of 'not Release' in the warning

Add CMake warning when building without OpenMP

Fix test compilation errors in GCC 9

Add OctoMap export functions and tests

The export functions don't work correctly for multires pipelines yet.

Fix dependency documentation issues

- Add OpenCV dependency.
- Mention than googletest must be compiled from source and remove from
  script.

Update Doxygen configuration

Allows showing namespace-level functions.

+++ se_apps/include/interface.h
@@ -34,7 +34,7 @@ enum ReaderType {
 
 struct ReaderConfiguration {
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  int fps;
+  float fps;
   bool blocking_read;
   std::string data_path;
   std::string groundtruth_path;
@@ -84,7 +84,7 @@ class DepthReader {
       //			fflush(stdout);
       //		}
 
-      if (_fps == 0) {
+      if (_fps == 0.0f) {
         _frame++;
         return;
       }
@@ -158,7 +158,7 @@ class DepthReader {
   protected:
     int _frame;
     size_t _pose_num;
-    int _fps;
+    float _fps;
     bool _blocking_read;
     std::string _data_path;
     std::string _groundtruth_path;
@@ -188,7 +188,7 @@ class SceneDepthReader: public DepthReader {
     SceneDepthReader(const ReaderConfiguration& config)
       : SceneDepthReader(config.data_path, config.fps, config.blocking_read){ }
 
-    SceneDepthReader(std::string dir, int fps, bool blocking_read) :
+    SceneDepthReader(std::string dir, float fps, bool blocking_read) :
       DepthReader(), _dir(dir), _inSize(make_uint2(640, 480)) {
         std::cerr << "No such directory " << dir << std::endl;
         struct stat st;
@@ -333,7 +333,7 @@ class RawDepthReader: public DepthReader {
      *
      * @deprecated Might be removed in the future.
      */
-    RawDepthReader(std::string filename, int fps, bool blocking_read) :
+    RawDepthReader(std::string filename, float fps, bool blocking_read) :
       DepthReader(), _rawFilePtr(fopen(filename.c_str(), "rb")) {
 
         size_t res = fread(&(_inSize), sizeof(_inSize), 1, _rawFilePtr);
@@ -577,7 +577,7 @@ class OpenNIDepthReader: public DepthReader {
     OpenNIDepthReader(const ReaderConfiguration& config)
       : OpenNIDepthReader(config.data_path, config.fps, config.blocking_read){ }
 
-    OpenNIDepthReader(std::string filename, int fps, bool blocking_read) :
+    OpenNIDepthReader(std::string filename, float fps, bool blocking_read) :
       DepthReader(), _pFile(fopen(filename.c_str(), "rb")) {
 
         cameraActive = false;
@@ -635,7 +635,7 @@ class OpenNIDepthReader: public DepthReader {
         if(!device.isFile())
         {
           openni::VideoMode newDepthMode;
-          newDepthMode.setFps(fps == 0 ? 30 : fps);
+          newDepthMode.setFps(fps == 0.0f ? 30 : fps);
           newDepthMode.setPixelFormat(openni::PIXEL_FORMAT_DEPTH_1_MM);
           newDepthMode.setResolution(640,480);
 
@@ -648,7 +648,7 @@ class OpenNIDepthReader: public DepthReader {
           }
 
           openni::VideoMode newRGBMode;
-          newRGBMode.setFps(fps == 0 ? 30 : fps);
+          newRGBMode.setFps(fps == 0.0f ? 30 : fps);
           newRGBMode.setPixelFormat(openni::PIXEL_FORMAT_RGB888);
           newRGBMode.setResolution(640,480);
